Rename packages to Java standard. - Also, adhere to Google's style guidelines.
