+++ app/src/main/java/com/cornellappdev/android/eatery/ExpandableListAdapter.java
@@ -9,8 +9,8 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.BaseExpandableListAdapter;
 import android.widget.TextView;
-import com.cornellappdev.android.eatery.Model.CafeteriaModel;
-import com.cornellappdev.android.eatery.Model.MealModel;
+import com.cornellappdev.android.eatery.model.CafeteriaModel;
+import com.cornellappdev.android.eatery.model.MealModel;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Date;
