+++ app/src/main/java/com/cornellappdev/android/eatery/MenuActivity.java
@@ -23,7 +23,7 @@ import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.TextView;
-import com.cornellappdev.android.eatery.Model.CafeteriaModel;
+import com.cornellappdev.android.eatery.model.CafeteriaModel;
 import com.facebook.drawee.view.SimpleDraweeView;
 import java.util.ArrayList;
