+++ app/src/main/java/com/cornellappdev/android/eatery/WeeklyMenuActivity.java
@@ -16,8 +16,8 @@ import android.view.View;
 import android.widget.LinearLayout;
 import android.widget.ScrollView;
 import android.widget.TextView;
-import com.cornellappdev.android.eatery.Model.CafeteriaModel;
-import com.cornellappdev.android.eatery.Model.MealModel;
+import com.cornellappdev.android.eatery.model.CafeteriaModel;
+import com.cornellappdev.android.eatery.model.MealModel;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Calendar;
